api.webchat/requirements.txt
asgiref==3.8.1
bidict==0.23.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
dnspython==2.6.1
eventlet==0.37.0
greenlet==3.1.1
h11==0.14.0
Markdown==3.7
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
simple-websocket==1.0.0
sqlparse==0.5.1
wsproto==1.2.0
